refactor(host): create claims and jetstream modules
