+++ crates/host/src/wasmbus/mod.rs
@@ -20,15 +20,17 @@ use std::task::{Context, Poll};
 use std::time::Duration;
 
 use anyhow::{anyhow, bail, ensure, Context as _};
-use async_nats::jetstream::kv::{Entry as KvEntry, Operation, Store};
+use async_nats::jetstream::kv::Store;
 use base64::engine::general_purpose::STANDARD;
 use base64::Engine;
 use bytes::{BufMut, Bytes, BytesMut};
+use claims::{Claims, StoredClaims};
 use cloudevents::{EventBuilder, EventBuilderV10};
 use futures::future::Either;
 use futures::stream::{AbortHandle, Abortable, SelectAll};
 use futures::{join, stream, try_join, Stream, StreamExt, TryFutureExt, TryStreamExt};
 use hyper_util::rt::{TokioExecutor, TokioIo};
+use jetstream::{create_bucket, ComponentSpecification};
 use nkeys::{KeyPair, KeyPairType, XKey};
 use secrecy::Secret;
 use serde::{Deserialize, Serialize};
@@ -43,7 +45,7 @@ use tokio_stream::wrappers::IntervalStream;
 use tracing::{debug, debug_span, error, info, instrument, trace, warn, Instrument as _};
 use tracing_opentelemetry::OpenTelemetrySpanExt;
 use uuid::Uuid;
-use wascap::{jwt, prelude::ClaimsBuilder};
+use wascap::jwt;
 use wasmcloud_control_interface::{
     ComponentAuctionAck, ComponentAuctionRequest, ComponentDescription, CtlResponse,
     DeleteInterfaceLinkDefinitionRequest, HostInventory, HostLabel, HostLabelIdentifier, Link,
@@ -68,9 +70,11 @@ use crate::{
     RegistryAuth, RegistryConfig, RegistryType, ResourceRef, SecretsManager,
 };
 
+mod claims;
 mod event;
 mod experimental;
 mod handler;
+mod jetstream;
 mod providers;
 
 pub mod config;
@@ -413,127 +417,6 @@ pub struct Host {
     /// A set of host tasks
     #[allow(unused)]
     tasks: JoinSet<()>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Default)]
-/// The specification of a component that is or did run in the lattice. This contains all of the information necessary to
-/// instantiate a component in the lattice (url and digest) as well as configuration and links in order to facilitate
-/// runtime execution of the component. Each `import` in a component's WIT world will need a corresponding link for the
-/// host runtime to route messages to the correct component.
-pub struct ComponentSpecification {
-    /// The URL of the component, file, OCI, or otherwise
-    url: String,
-    /// All outbound links from this component to other components, used for routing when calling a component `import`
-    links: Vec<Link>,
-    ////
-    // Possible additions in the future, left in as comments to facilitate discussion
-    ////
-    // /// The claims embedded in the component, if present
-    // claims: Option<Claims>,
-    // /// SHA256 digest of the component, used for checking uniqueness of component IDs
-    // digest: String
-    // /// (Advanced) Additional routing topics to subscribe on in addition to the component ID.
-    // routing_groups: Vec<String>,
-}
-
-impl ComponentSpecification {
-    /// Create a new empty component specification with the given ID and URL
-    pub fn new(url: impl AsRef<str>) -> Self {
-        Self {
-            url: url.as_ref().to_string(),
-            links: Vec::new(),
-        }
-    }
-}
-
-#[allow(clippy::large_enum_variant)] // Without this clippy complains component is at least 0 bytes while provider is at least 280 bytes. That doesn't make sense
-enum Claims {
-    Component(jwt::Claims<jwt::Component>),
-    Provider(jwt::Claims<jwt::CapabilityProvider>),
-}
-
-impl Claims {
-    fn subject(&self) -> &str {
-        match self {
-            Claims::Component(claims) => &claims.subject,
-            Claims::Provider(claims) => &claims.subject,
-        }
-    }
-}
-
-impl From<StoredClaims> for Claims {
-    fn from(claims: StoredClaims) -> Self {
-        match claims {
-            StoredClaims::Component(claims) => {
-                let name = (!claims.name.is_empty()).then_some(claims.name);
-                let rev = claims.revision.parse().ok();
-                let ver = (!claims.version.is_empty()).then_some(claims.version);
-                let tags = (!claims.tags.is_empty()).then_some(claims.tags);
-                let call_alias = (!claims.call_alias.is_empty()).then_some(claims.call_alias);
-                let metadata = jwt::Component {
-                    name,
-                    tags,
-                    rev,
-                    ver,
-                    call_alias,
-                    ..Default::default()
-                };
-                let claims = ClaimsBuilder::new()
-                    .subject(&claims.subject)
-                    .issuer(&claims.issuer)
-                    .with_metadata(metadata)
-                    .build();
-                Claims::Component(claims)
-            }
-            StoredClaims::Provider(claims) => {
-                let name = (!claims.name.is_empty()).then_some(claims.name);
-                let rev = claims.revision.parse().ok();
-                let ver = (!claims.version.is_empty()).then_some(claims.version);
-                let config_schema: Option<serde_json::Value> = claims
-                    .config_schema
-                    .and_then(|schema| serde_json::from_str(&schema).ok());
-                let metadata = jwt::CapabilityProvider {
-                    name,
-                    rev,
-                    ver,
-                    config_schema,
-                    ..Default::default()
-                };
-                let claims = ClaimsBuilder::new()
-                    .subject(&claims.subject)
-                    .issuer(&claims.issuer)
-                    .with_metadata(metadata)
-                    .build();
-                Claims::Provider(claims)
-            }
-        }
-    }
-}
-
-#[instrument(level = "debug", skip_all)]
-async fn create_bucket(
-    jetstream: &async_nats::jetstream::Context,
-    bucket: &str,
-) -> anyhow::Result<Store> {
-    // Don't create the bucket if it already exists
-    if let Ok(store) = jetstream.get_key_value(bucket).await {
-        info!(%bucket, "bucket already exists. Skipping creation.");
-        return Ok(store);
-    }
-
-    match jetstream
-        .create_key_value(async_nats::jetstream::kv::Config {
-            bucket: bucket.to_string(),
-            ..Default::default()
-        })
-        .await
-    {
-        Ok(store) => {
-            info!(%bucket, "created bucket with 1 replica");
-            Ok(store)
-        }
-        Err(err) => Err(anyhow!(err).context(format!("failed to create bucket '{bucket}'"))),
-    }
 }
 
 /// Given the NATS address, authentication jwt, seed, tls requirement and optional request timeout,
@@ -3516,264 +3399,6 @@ impl Host {
             .context("failed to publish provider link definition delete")
     }
 
-    /// Retrieve a component specification based on the provided ID. The outer Result is for errors
-    /// accessing the store, and the inner option indicates if the spec exists.
-    #[instrument(level = "debug", skip_all)]
-    async fn get_component_spec(&self, id: &str) -> anyhow::Result<Option<ComponentSpecification>> {
-        let key = format!("COMPONENT_{id}");
-        let spec = self
-            .data
-            .get(key)
-            .await
-            .context("failed to get component spec")?
-            .map(|spec_bytes| serde_json::from_slice(&spec_bytes))
-            .transpose()
-            .context(format!(
-                "failed to deserialize stored component specification for {id}"
-            ))?;
-        Ok(spec)
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    async fn store_component_spec(
-        &self,
-        id: impl AsRef<str>,
-        spec: &ComponentSpecification,
-    ) -> anyhow::Result<()> {
-        let id = id.as_ref();
-        let key = format!("COMPONENT_{id}");
-        let bytes = serde_json::to_vec(spec)
-            .context("failed to serialize component spec")?
-            .into();
-        self.data
-            .put(key, bytes)
-            .await
-            .context("failed to put component spec")?;
-        Ok(())
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    async fn store_claims(&self, claims: Claims) -> anyhow::Result<()> {
-        match &claims {
-            Claims::Component(claims) => {
-                self.store_component_claims(claims.clone()).await?;
-            }
-            Claims::Provider(claims) => {
-                let mut provider_claims = self.provider_claims.write().await;
-                provider_claims.insert(claims.subject.clone(), claims.clone());
-            }
-        };
-        let claims: StoredClaims = claims.try_into()?;
-        let subject = match &claims {
-            StoredClaims::Component(claims) => &claims.subject,
-            StoredClaims::Provider(claims) => &claims.subject,
-        };
-        let key = format!("CLAIMS_{subject}");
-        trace!(?claims, ?key, "storing claims");
-
-        let bytes = serde_json::to_vec(&claims)
-            .context("failed to serialize claims")?
-            .into();
-        self.data
-            .put(key, bytes)
-            .await
-            .context("failed to put claims")?;
-        Ok(())
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    async fn process_component_spec_put(
-        &self,
-        id: impl AsRef<str>,
-        value: impl AsRef<[u8]>,
-        _publish: bool,
-    ) -> anyhow::Result<()> {
-        let id = id.as_ref();
-        debug!(id, "process component spec put");
-
-        let spec: ComponentSpecification = serde_json::from_slice(value.as_ref())
-            .context("failed to deserialize component specification")?;
-
-        // Compute all new links that do not exist in the host map, which we'll use to
-        // publish to any running providers that are the source or target of the link.
-        // Computing this ahead of time is a tradeoff to hold only one lock at the cost of
-        // allocating an extra Vec. This may be a good place to optimize allocations.
-        let new_links = {
-            let all_links = self.links.read().await;
-            spec.links
-                .iter()
-                .filter(|spec_link| {
-                    // Retain only links that do not exist in the host map
-                    !all_links
-                        .iter()
-                        .filter_map(|(source_id, links)| {
-                            // Only consider links that are either the source or target of the new link
-                            if source_id == spec_link.source_id() || source_id == spec_link.target()
-                            {
-                                Some(links)
-                            } else {
-                                None
-                            }
-                        })
-                        .flatten()
-                        .any(|host_link| *spec_link == host_link)
-                })
-                .collect::<Vec<_>>()
-        };
-
-        {
-            // Acquire lock once in this block to avoid continually trying to acquire it.
-            let providers = self.providers.read().await;
-            // For every new link, if a provider is running on this host as the source or target,
-            // send the link to the provider for handling based on the xkey public key.
-            for link in new_links {
-                if let Some(provider) = providers.get(link.source_id()) {
-                    if let Err(e) = self.put_provider_link(provider, link).await {
-                        error!(?e, "failed to put provider link");
-                    }
-                }
-                if let Some(provider) = providers.get(link.target()) {
-                    if let Err(e) = self.put_provider_link(provider, link).await {
-                        error!(?e, "failed to put provider link");
-                    }
-                }
-            }
-        }
-
-        // If the component is already running, update the links
-        if let Some(component) = self.components.write().await.get(id) {
-            *component.handler.instance_links.write().await = component_import_links(&spec.links);
-            // NOTE(brooksmtownsend): We can consider updating the component if the image URL changes
-        };
-
-        // Insert the links into host map
-        self.links.write().await.insert(id.to_string(), spec.links);
-
-        Ok(())
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    async fn process_component_spec_delete(
-        &self,
-        id: impl AsRef<str>,
-        _value: impl AsRef<[u8]>,
-        _publish: bool,
-    ) -> anyhow::Result<()> {
-        let id = id.as_ref();
-        debug!(id, "process component delete");
-        // TODO: TBD: stop component if spec deleted?
-        if self.components.write().await.get(id).is_some() {
-            warn!(
-                component_id = id,
-                "component spec deleted, but component is still running"
-            );
-        }
-        Ok(())
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    async fn process_claims_put(
-        &self,
-        pubkey: impl AsRef<str>,
-        value: impl AsRef<[u8]>,
-    ) -> anyhow::Result<()> {
-        let pubkey = pubkey.as_ref();
-
-        debug!(pubkey, "process claim entry put");
-
-        let stored_claims: StoredClaims =
-            serde_json::from_slice(value.as_ref()).context("failed to decode stored claims")?;
-        let claims = Claims::from(stored_claims);
-
-        ensure!(claims.subject() == pubkey, "subject mismatch");
-        match claims {
-            Claims::Component(claims) => self.store_component_claims(claims).await,
-            Claims::Provider(claims) => {
-                let mut provider_claims = self.provider_claims.write().await;
-                provider_claims.insert(claims.subject.clone(), claims);
-                Ok(())
-            }
-        }
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    async fn process_claims_delete(
-        &self,
-        pubkey: impl AsRef<str>,
-        value: impl AsRef<[u8]>,
-    ) -> anyhow::Result<()> {
-        let pubkey = pubkey.as_ref();
-
-        debug!(pubkey, "process claim entry deletion");
-
-        let stored_claims: StoredClaims =
-            serde_json::from_slice(value.as_ref()).context("failed to decode stored claims")?;
-        let claims = Claims::from(stored_claims);
-
-        ensure!(claims.subject() == pubkey, "subject mismatch");
-
-        match claims {
-            Claims::Component(claims) => {
-                let mut component_claims = self.component_claims.write().await;
-                component_claims.remove(&claims.subject);
-            }
-            Claims::Provider(claims) => {
-                let mut provider_claims = self.provider_claims.write().await;
-                provider_claims.remove(&claims.subject);
-            }
-        }
-
-        Ok(())
-    }
-
-    #[instrument(level = "trace", skip_all)]
-    async fn process_entry(
-        &self,
-        KvEntry {
-            key,
-            value,
-            operation,
-            ..
-        }: KvEntry,
-        publish: bool,
-    ) {
-        let key_id = key.split_once('_');
-        let res = match (operation, key_id) {
-            (Operation::Put, Some(("COMPONENT", id))) => {
-                self.process_component_spec_put(id, value, publish).await
-            }
-            (Operation::Delete, Some(("COMPONENT", id))) => {
-                self.process_component_spec_delete(id, value, publish).await
-            }
-            (Operation::Put, Some(("LINKDEF", _id))) => {
-                debug!("ignoring deprecated LINKDEF put operation");
-                Ok(())
-            }
-            (Operation::Delete, Some(("LINKDEF", _id))) => {
-                debug!("ignoring deprecated LINKDEF delete operation");
-                Ok(())
-            }
-            (Operation::Put, Some(("CLAIMS", pubkey))) => {
-                self.process_claims_put(pubkey, value).await
-            }
-            (Operation::Delete, Some(("CLAIMS", pubkey))) => {
-                self.process_claims_delete(pubkey, value).await
-            }
-            (operation, Some(("REFMAP", id))) => {
-                // TODO: process REFMAP entries
-                debug!(?operation, id, "ignoring REFMAP entry");
-                Ok(())
-            }
-            _ => {
-                warn!(key, ?operation, "unsupported KV bucket entry");
-                Ok(())
-            }
-        };
-        if let Err(error) = &res {
-            error!(key, ?operation, ?error, "failed to process KV bucket entry");
-        }
-    }
-
     async fn fetch_config_and_secrets(
         &self,
         config_names: &[String],
@@ -3963,240 +3588,6 @@ fn serialize_ctl_response<T: Serialize>(
     ctl_response: Option<CtlResponse<T>>,
 ) -> Option<anyhow::Result<Vec<u8>>> {
     ctl_response.map(|resp| serde_json::to_vec(&resp).map_err(anyhow::Error::from))
-}
-
-// TODO: remove StoredClaims in #1093
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(untagged)]
-enum StoredClaims {
-    Component(StoredComponentClaims),
-    Provider(StoredProviderClaims),
-}
-
-#[derive(Debug, Default, Serialize, Deserialize)]
-struct StoredComponentClaims {
-    call_alias: String,
-    #[serde(alias = "iss")]
-    issuer: String,
-    name: String,
-    #[serde(alias = "rev")]
-    revision: String,
-    #[serde(alias = "sub")]
-    subject: String,
-    #[serde(deserialize_with = "deserialize_messy_vec")]
-    tags: Vec<String>,
-    version: String,
-}
-
-#[derive(Debug, Default, Serialize, Deserialize)]
-struct StoredProviderClaims {
-    #[serde(alias = "iss")]
-    issuer: String,
-    name: String,
-    #[serde(alias = "rev")]
-    revision: String,
-    #[serde(alias = "sub")]
-    subject: String,
-    version: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    config_schema: Option<String>,
-}
-
-impl TryFrom<Claims> for StoredClaims {
-    type Error = anyhow::Error;
-
-    fn try_from(claims: Claims) -> Result<Self, Self::Error> {
-        match claims {
-            Claims::Component(jwt::Claims {
-                issuer,
-                subject,
-                metadata,
-                ..
-            }) => {
-                let jwt::Component {
-                    name,
-                    tags,
-                    rev,
-                    ver,
-                    call_alias,
-                    ..
-                } = metadata.context("no metadata found on component claims")?;
-                Ok(StoredClaims::Component(StoredComponentClaims {
-                    call_alias: call_alias.unwrap_or_default(),
-                    issuer,
-                    name: name.unwrap_or_default(),
-                    revision: rev.unwrap_or_default().to_string(),
-                    subject,
-                    tags: tags.unwrap_or_default(),
-                    version: ver.unwrap_or_default(),
-                }))
-            }
-            Claims::Provider(jwt::Claims {
-                issuer,
-                subject,
-                metadata,
-                ..
-            }) => {
-                let jwt::CapabilityProvider {
-                    name,
-                    rev,
-                    ver,
-                    config_schema,
-                    ..
-                } = metadata.context("no metadata found on provider claims")?;
-                Ok(StoredClaims::Provider(StoredProviderClaims {
-                    issuer,
-                    name: name.unwrap_or_default(),
-                    revision: rev.unwrap_or_default().to_string(),
-                    subject,
-                    version: ver.unwrap_or_default(),
-                    config_schema: config_schema.map(|schema| schema.to_string()),
-                }))
-            }
-        }
-    }
-}
-
-impl TryFrom<&Claims> for StoredClaims {
-    type Error = anyhow::Error;
-
-    fn try_from(claims: &Claims) -> Result<Self, Self::Error> {
-        match claims {
-            Claims::Component(jwt::Claims {
-                issuer,
-                subject,
-                metadata,
-                ..
-            }) => {
-                let jwt::Component {
-                    name,
-                    tags,
-                    rev,
-                    ver,
-                    call_alias,
-                    ..
-                } = metadata
-                    .as_ref()
-                    .context("no metadata found on component claims")?;
-                Ok(StoredClaims::Component(StoredComponentClaims {
-                    call_alias: call_alias.clone().unwrap_or_default(),
-                    issuer: issuer.clone(),
-                    name: name.clone().unwrap_or_default(),
-                    revision: rev.unwrap_or_default().to_string(),
-                    subject: subject.clone(),
-                    tags: tags.clone().unwrap_or_default(),
-                    version: ver.clone().unwrap_or_default(),
-                }))
-            }
-            Claims::Provider(jwt::Claims {
-                issuer,
-                subject,
-                metadata,
-                ..
-            }) => {
-                let jwt::CapabilityProvider {
-                    name,
-                    rev,
-                    ver,
-                    config_schema,
-                    ..
-                } = metadata
-                    .as_ref()
-                    .context("no metadata found on provider claims")?;
-                Ok(StoredClaims::Provider(StoredProviderClaims {
-                    issuer: issuer.clone(),
-                    name: name.clone().unwrap_or_default(),
-                    revision: rev.unwrap_or_default().to_string(),
-                    subject: subject.clone(),
-                    version: ver.clone().unwrap_or_default(),
-                    config_schema: config_schema.as_ref().map(ToString::to_string),
-                }))
-            }
-        }
-    }
-}
-
-#[allow(clippy::implicit_hasher)]
-impl From<StoredClaims> for HashMap<String, String> {
-    fn from(claims: StoredClaims) -> Self {
-        match claims {
-            StoredClaims::Component(claims) => HashMap::from([
-                ("call_alias".to_string(), claims.call_alias),
-                ("iss".to_string(), claims.issuer.clone()), // TODO: remove in #1093
-                ("issuer".to_string(), claims.issuer),
-                ("name".to_string(), claims.name),
-                ("rev".to_string(), claims.revision.clone()), // TODO: remove in #1093
-                ("revision".to_string(), claims.revision),
-                ("sub".to_string(), claims.subject.clone()), // TODO: remove in #1093
-                ("subject".to_string(), claims.subject),
-                ("tags".to_string(), claims.tags.join(",")),
-                ("version".to_string(), claims.version),
-            ]),
-            StoredClaims::Provider(claims) => HashMap::from([
-                ("iss".to_string(), claims.issuer.clone()), // TODO: remove in #1093
-                ("issuer".to_string(), claims.issuer),
-                ("name".to_string(), claims.name),
-                ("rev".to_string(), claims.revision.clone()), // TODO: remove in #1093
-                ("revision".to_string(), claims.revision),
-                ("sub".to_string(), claims.subject.clone()), // TODO: remove in #1093
-                ("subject".to_string(), claims.subject),
-                ("version".to_string(), claims.version),
-                (
-                    "config_schema".to_string(),
-                    claims.config_schema.unwrap_or_default(),
-                ),
-            ]),
-        }
-    }
-}
-
-fn deserialize_messy_vec<'de, D: serde::Deserializer<'de>>(
-    deserializer: D,
-) -> Result<Vec<String>, D::Error> {
-    MessyVec::deserialize(deserializer).map(|messy_vec| messy_vec.0)
-}
-// Helper struct to deserialize either a comma-delimited string or an actual array of strings
-struct MessyVec(pub Vec<String>);
-
-struct MessyVecVisitor;
-
-// Since this is "temporary" code to preserve backwards compatibility with already-serialized claims,
-// we use fully-qualified names instead of importing
-impl<'de> serde::de::Visitor<'de> for MessyVecVisitor {
-    type Value = MessyVec;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-        formatter.write_str("string or array of strings")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut values = Vec::new();
-
-        while let Some(value) = seq.next_element()? {
-            values.push(value);
-        }
-
-        Ok(MessyVec(values))
-    }
-
-    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-    where
-        E: serde::de::Error,
-    {
-        Ok(MessyVec(value.split(',').map(String::from).collect()))
-    }
-}
-
-impl<'de> Deserialize<'de> for MessyVec {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::de::Deserializer<'de>,
-    {
-        deserializer.deserialize_any(MessyVecVisitor)
-    }
 }
 
 fn human_friendly_uptime(uptime: Duration) -> String {
